requirements.txt
azure-cosmos
pandas